Bug fixes for User Ranking entities. Logging for Reporting Job. Fix for precision to 2 decimal places for rating.

+++ Source/SSCore/src/main/java/com/realtech/socialsurvey/core/entities/UserRankingPastYearMain.java
@@ -68,7 +68,7 @@ public class UserRankingPastYearMain
     private int totalReviews;
     
     @Column(name = "average_rating")
-    private int averageRating;
+    private float averageRating;
     
     @Column(name = "is_eligible")
     private	int isEligible;
@@ -79,12 +79,11 @@ public class UserRankingPastYearMain
     @Transient
     private String profileImageUrlThumbnail;
     
-	public int getAverageRating() {
+	public float getAverageRating() {
 		return averageRating;
 	}
 
-
-	public void setAverageRating(int averageRating) {
+	public void setAverageRating(float averageRating) {
 		this.averageRating = averageRating;
 	}
