+++ Source/SSCore/src/main/java/com/realtech/socialsurvey/core/entities/UserRankingPastYearsMain.java
@@ -64,7 +64,7 @@ public class UserRankingPastYearsMain
     private int totalReviews;
     
     @Column(name = "average_rating")
-    private int averageRating;
+    private float averageRating;
     
     @Column(name = "is_eligible")
     private	int isEligible;
@@ -216,17 +216,15 @@ public class UserRankingPastYearsMain
         this.totalReviews = totalReviews;
     }
 
-    public int getAverageRating()
-    {
-        return averageRating;
-    }
+    public float getAverageRating() {
+		return averageRating;
+	}
 
-    public void setAverageRating( int averageRating )
-    {
-        this.averageRating = averageRating;
-    }
+	public void setAverageRating(float averageRating) {
+		this.averageRating = averageRating;
+	}
 
-    public int getIsEligible() {
+	public int getIsEligible() {
 		return isEligible;
 	}
