+++ Source/SSCore/src/main/java/com/realtech/socialsurvey/core/entities/UserRankingThisMonthRegion.java
@@ -72,7 +72,7 @@ public class UserRankingThisMonthRegion
     private int totalReviews;
     
     @Column(name = "average_rating")
-    private int averageRating;
+    private float averageRating;
     
     @Column(name = "is_eligible")
     private	int isEligible;
@@ -86,11 +86,11 @@ public class UserRankingThisMonthRegion
     @Transient
     private String profileImageUrlThumbnail;
     
-	public int getAverageRating() {
+	public float getAverageRating() {
 		return averageRating;
 	}
 
-	public void setAverageRating(int averageRating) {
+	public void setAverageRating(float averageRating) {
 		this.averageRating = averageRating;
 	}
