+++ Source/SSCore/src/main/java/com/realtech/socialsurvey/core/entities/UserRankingThisYearMain.java
@@ -69,7 +69,7 @@ public class UserRankingThisYearMain
     private int totalReviews;
     
     @Column(name = "average_rating")
-    private int averageRating;
+    private float averageRating;
     
     @Column(name = "is_eligible")
     private	int isEligible;
@@ -80,11 +80,11 @@ public class UserRankingThisYearMain
     @Transient
     private String profileImageUrlThumbnail;
     
-	public int getAverageRating() {
+    public float getAverageRating() {
 		return averageRating;
 	}
 
-	public void setAverageRating(int averageRating) {
+	public void setAverageRating(float averageRating) {
 		this.averageRating = averageRating;
 	}
